Bug fixes
Several Bug fixes:
Top Destinations now work correctly when viewing bike stations.
Added several stations to the database that were mentioned in trips but were not in the list of bike stations.
Merged "Add Journey" and "Add Station" tabs into one.

New feature:
When viewing bike stations in the list, you can click on them to move to the page displaying additional details about the chosen station.
Added a loading spinner to the station details page, as the page can take a long time to load on azure cloud.